Removed old right click logic. Closes #51
There's no longer a way for me to see if the interaction was triggered
by a right or left click. Neither can I intercept a right click, since
the first sport where I could change this, it's not possible for me to
see what triggered it in the first place, making it equal to a normal
left click and will thus be treated as one from now on.